+++ packages/neuron-ui/src/components/VerifyHardwareAddress/index.tsx
@@ -17,6 +17,8 @@ import {
 import { ErrorCode, errorFormatter, isSuccessResponse, useDidMount } from 'utils'
 import { CkbAppNotFoundException, DeviceNotFoundException } from 'exceptions'
 import CopyZone from 'widgets/CopyZone'
+import { ADDRESS_LENGTH } from 'utils/const'
+import { AddressPrefix, addressToScript, scriptToAddress } from '@nervosnetwork/ckb-sdk-utils'
 import styles from './verifyHardwareAddress.module.scss'
 import VerifyError from './verify-error'
 
@@ -24,6 +26,26 @@ export interface VerifyHardwareAddressProps {
   address: string
   wallet: State.WalletIdentity
   onDismiss: () => void
+}
+
+const toLongAddr = (addr: string) => {
+  try {
+    const script = addressToScript(addr)
+    const isMainnet = addr.startsWith(AddressPrefix.Mainnet)
+    return scriptToAddress(script, isMainnet)
+  } catch {
+    return ''
+  }
+}
+
+const verifyAddressEqual = (address: string, compared?: string) => {
+  if (!compared) {
+    return false
+  }
+  if (address.length !== compared.length) {
+    return toLongAddr(address) === toLongAddr(compared)
+  }
+  return address === compared
 }
 
 const VerifyHardwareAddress = ({ address, wallet, onDismiss }: VerifyHardwareAddressProps) => {
@@ -139,7 +161,7 @@ const VerifyHardwareAddress = ({ address, wallet, onDismiss }: VerifyHardwareAdd
     const res = await getDevicePublicKey()
     if (isSuccessResponse(res)) {
       const { result } = res
-      if (result?.address === address) {
+      if (verifyAddressEqual(address, result?.address)) {
         setStatus(verifiedStatus)
       } else {
         setStatus(invalidStatus)
@@ -176,9 +198,17 @@ const VerifyHardwareAddress = ({ address, wallet, onDismiss }: VerifyHardwareAdd
                 <CopyZone
                   content={address}
                   name={t('hardware-verify-address.actions.copy-address')}
-                  style={{ lineHeight: '1.625rem' }}
+                  style={{ lineHeight: '1.625rem', width: '400px', display: 'flex' }}
                 >
-                  {address}
+                  {address.length > ADDRESS_LENGTH ? (
+                    <>
+                      <span className={styles.addressOverflow}>{address.slice(0, -21)}</span>
+                      <span>...</span>
+                      <span>{address.slice(-20)}</span>
+                    </>
+                  ) : (
+                    address
+                  )}
                 </CopyZone>
               </td>
             </tr>