+++ packages/neuron-wallet/src/controllers/app/index.ts
@@ -15,6 +15,7 @@ import SyncApiController from 'controllers/sync-api'
 import { SETTINGS_WINDOW_TITLE } from 'utils/const'
 import IndexerService from 'services/indexer'
 import { stopCkbNode } from 'services/ckb-runner'
+import startMonitor from 'services/monitor'
 
 const app = electronApp
 
@@ -50,6 +51,8 @@ export default class AppController {
     SyncApiController.getInstance().mount()
 
     await this.openWindow()
+
+    startMonitor()
   }
 
   /**